Debugging

+++ suspend.sh
@@ -21,51 +21,60 @@ LOCATION=`cat ${SNOOP_DIR}/.DeviceLoc`
 touch /tmp/Snoopy/STOP_SNIFFING
 
 # Give sub-processes a chance to clean things up...
-sleep 1m 
+sleep 1m
 
-sudo kill -KILL $(cat /tmp/Snoopy/Airodump.pid)
+if [ -f /tmp/Snoopy/Airodump.pid ]; then
+    sudo kill -s KILL $(cat /tmp/Snoopy/Airodump.pid)
+fi
+
+if [ -f /tmp/Snoopy/Snoopy.pid ]; then
+    sudo kill $(cat /tmp/Snoopy/Snoopy.pid)
+fi
 
 # Any straggling processes:
-SNOOP=$(ps -aux | grep snoopy   | grep -v grep | awk '{print $2}' | sed ':a;N;$!ba;s/\n/ /g');
+# SNOOP=$(ps -aux | grep snoopy   | grep -v grep | awk '{print $2}' | sed ':a;N;$!ba;s/\n/ /g');
 AIRNG=$(ps -aux | grep airodump | grep -v grep | awk '{print $2}' | sed ':a;N;$!ba;s/\n/ /g');
-sudo kill -s KILL $AIRNG $SNOOP
+sudo kill -s KILL $AIRNG
 
 sudo airmon-ng stop $(ifconfig -a | sed 's/[ \t].*//;/^$/d' | grep mon);
 
 IFACE=$(ifconfig -a | sed 's/[ \t].*//;/^$/d' | grep wlan);
-sudo ifconfig $IFACE up;
 
-sudo iwconfig $IFACE mode managed;
-# sudo iwconfig $IFACE essid "${NETWORK}";
-sudo dhclient $IFACE;
+if [ -f "$DATABASE" ]; then
+    sudo ifconfig $IFACE up;
 
-# 'COUNTER' value is overset only during testing. (Should be reduced to '1' for final deployment.)
-let COUNTER=7;
-# let COUNTER=1;
+    sudo iwconfig $IFACE mode managed;
+    # sudo iwconfig $IFACE essid "${NETWORK}";
+    sudo dhclient $IFACE;
 
-while [  $COUNTER -lt 4 ]; do
-    if [ scp $DATABASE "${USER}@${SERVER}:/home/${USER}/${LOCATION}/${DEVICE}" -eq 0 ]; then
-        rm $DATABASE;
-        let COUNTER=10;
+    # 'COUNTER' value is overset only during testing. (Should be reduced to '1' for final deployment.)
+    let COUNTER=7;
+    # let COUNTER=1;
+
+    while [  $COUNTER -lt 4 ]; do
+        if [ scp $DATABASE "${USER}@${SERVER}:/home/${USER}/${LOCATION}/${DEVICE}" -eq 0 ]; then
+            rm $DATABASE;
+            let COUNTER=10;
+        else
+            sleep 30;
+            let COUNTER=COUNTER+1;
+        fi
+    done
+
+    if [ $COUNTER -eq 10 ]; then
+        echo "Database synced successfully." | tee -a ./Database.log
     else
-        sleep 30;
-        let COUNTER=COUNTER+1;
+        filename=$(basename "$DATABASE");
+        ext="${filename##*.}";
+        filename="${filename%.*}";
+
+        NOW=$(date +%F@%T);
+        echo "[${NOW}] :: Database failed to sync. Data will be maintained locally." | tee -a ./Database.log
+        mv $DATABASE "$(dirname "$DATABASE")/${filename}_${NOW}.${ext}";
     fi
-done
 
-if [ $COUNTER -eq 10 ]; then
-    echo "Database synced successfully." | tee -a ./Database.log
-else
-    filename=$(basename "$DATABASE");
-    ext="${filename##*.}";
-    filename="${filename%.*}";
-
-    NOW=$(date +%F@%T);
-    echo "[${NOW}] :: Database failed to sync. Data will be maintained locally." | tee -a ./Database.log
-    mv $DATABASE "$(dirname "$DATABASE")/${filename}_${NOW}.${ext}";
+    sudo rm -f /tmp/Snoopy/*
 fi
-
-sudo rm -f /tmp/Snoopy/*
 
 sudo ifconfig $IFACE down;
